backend/requirements.txt
psycopg2-binary
fastapi
uvicorn
sqlalchemy
alembic
pydantic-settings
python-dotenv
requests
chromadb
apscheduler
python-pptx
trafilatura
sentence-transformers
pydantic
aiohttp
python-multipart
